requirements.txt
django==3.2.20
djangorestframework==3.14.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.8.6
pandas==2.1.1
openpyxl==3.1.2
gunicorn==20.0.4
